Common: Rename register{,_agent_events}.py
